% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{adjMat}
\alias{adjMat}
\title{Matrix Adjuster}
\usage{
adjMat(data, colName)
}
\arguments{
\item{data}{is data}

\item{colName}{is the column the user wants, in quotes

Uses dpylr on a matrix}
}
\description{
Matrix Adjuster
}
\examples{
data(flights)
adjMat(flights, "dep_time")
}
