% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModels.R
\name{buildModels}
\alias{buildModels}
\title{Builds models using locating arrays}
\usage{
buildModels(
  la_path,
  factor_data_path,
  response_dir,
  response_column,
  log_data,
  max_terms,
  models_n,
  new_models_n,
  logit
)
}
\arguments{
\item{la_path}{Path to locating array TSV.}

\item{response_dir}{Path to the response directory in TSV format.}

\item{response_column}{Response column as a string}

\item{log_data}{TRUE, FALSE Apply logarithm to response data}

\item{max_terms}{Number of terms in each model}

\item{models_n}{Number of models}

\item{new_models_n}{Number of models to be created at each iteration}

\item{logit}{TRUE, FALSE Create logistic regression models rather than linear}

\item{factor_data}{Path to factor data file in TSV format.}
}
\value{
A text output of all top models and iterations as well as summary statistics.
}
\description{
Builds models using locating arrays
}
\examples{
locatingArray_path = system.file("extdata", "programRepair_logit/LA.tsv", package="LATools")
factorData_path = system.file("extdata", "programRepair_logit/factor_data.tsv", package="LATools")
response_path = system.file("extdata", "programRepair_logit/Response", package="LATools")
##EXAMPLE 1: Use logisitc regression model on software repair data set.
buildModels(locatingArray_path, factorData_path, response_path, response_column = "better_than_original", logit = TRUE)
##EXAMPLE 2: Use linear regression model on network optimization data set.
locatingArray_path = system.file("extdata", "network_linear/LA.tsv", package="LATools")
factorData_path = system.file("extdata", "network_linear/factor_data.tsv", package="LATools")
response_path = system.file("extdata", "network_linear/Response", package="LATools")
buildModels(locatingArray_path, factorData_path, response_path, response_column = "MOS", logit = FALSE)
}
