% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gap}
\alias{gap}
\title{Example data from class}
\format{An object of class \code{data.frame} with 142 rows and 6 columns.}
\usage{
gap
}
\description{
Example data from class
}
\references{
\url{http://users.stat.umn.edu/~almquist/3811_examples/gapminder2007ex.csv}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
