% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{ggplotW}
\alias{ggplotW}
\title{GGPlot Wrapper}
\usage{
ggplotW(data, xcol, ycol, colorVar)
}
\arguments{
\item{data}{is data}

\item{xcol}{is the number of columns in x}

\item{ycol}{is the number of columns in y}

\item{colorVar}{is the color of the data points in the plot

Creates a graph}
}
\description{
GGPlot Wrapper
}
\examples{
data(flights)
ggplotW(flights, 2, 2, color = blue)
}
