\name{LATools-package}
\alias{LATools-package}
\alias{LATools}
\docType{package}
\title{
   Tool set for locating array model building and construction.
}
\description{
  This package is used to construct and analyze locating arrays. A locating array is a combinatorial tool used to design tests that include interactions. This tool set is able to construct locating arrays from a description of test factors. Once the locating array is constructed along and tested to produce results, an algorithm is used to build models. The model builder will inform the user of important factors and interactions in their system.
}
\details{


  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Leah Darwin

Maintainer: Leah Darwin <ljdarwin@asu.edu>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
