% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{logL}
\alias{logL}
\title{Log Likelihood}
\usage{
logL(x, fun, int)
}
\arguments{
\item{x}{is data}

\item{fun}{is a distribution function}

\item{int}{is the interval of x values

Calculates the log likelihood of a datset over a certain distribution and interval determined by the user}
}
\description{
Log Likelihood
}
\examples{
 
data(d)
logL(d, binomial, c(-100, 100))
}
