% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{myapply}
\alias{myapply}
\title{MyApply}
\usage{
myapply(X, MARGIN, FUN, ...)
}
\arguments{
\item{MARGIN}{is the dimiensions of the matrix.}

\item{FUN}{is a predefined function.

Returns a simplified array similar to other apply functiion already in R.}

\item{x}{is matrix}
}
\description{
MyApply
}
\examples{
data(d)
myapply(d, 1, mean)
}
