% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{spread}
\alias{spread}
\title{Measures of Spread}
\usage{
spread(d)
}
\arguments{
\item{x}{is a vector

Gives the mean, variance and standard deviation of a population from data, x.}
}
\description{
Measures of Spread
}
\examples{
data(d)
spread(d)
}
