% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcs.R
\name{spreadCheck}
\alias{spreadCheck}
\title{Measures of Spread w/ Checks}
\usage{
spreadCheck(d)
}
\arguments{
\item{x}{is a vector

Gives the mean, variance and standard deviation of a population from data, x in the form of a list. Has checks implemented to ensure the data given is proper}
}
\description{
Measures of Spread w/ Checks
}
\examples{
data(d) 
spread(d)
}
