% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModels.R
\name{buildModels}
\alias{buildModels}
\title{Builds models using locating arrays}
\usage{
buildModels(
  la_path,
  factor_data_path,
  response_dir,
  response_column,
  log_data,
  max_terms,
  models_n,
  new_models_n
)
}
\arguments{
\item{la_path}{Path to locating array TSV.}

\item{response_dir}{Path to the response directory in TSV format.}

\item{response_column}{Response column as a string}

\item{log_data}{TRUE, FALSE Apply logarithm to response data}

\item{max_terms}{Number of terms in each model}

\item{models_n}{Number of models}

\item{new_models_n}{Number of models to be created at each iteration}

\item{factor_data}{Path to factor data file in TSV format.}
}
\value{
A text output of all top models and iterations as well as summary statistics.
}
\description{
Builds models using locating arrays
}
\examples{
buildModels("LA/LA_LARGE.tsv", "FD/Factors_LARGE.tsv", "RE/responses_LARGE", "Throughput", "TRUE", 13, 50, 50)
}
