\name{LATools-package}
\alias{LATools-package}
\alias{LATools}
\docType{package}
\title{
   Tool set for locating array model building and construction.
}
\description{
  This package is used to construct and analyze locating arrays. A locating array is a combinatorial tool used to design tests that include interactions. This tool set is able to construct locating arrays from a description of test factors. Once the locating array is constructed along and tested to produce results, an algorithm is used to build models. The model builder will inform the user of important factors and interactions in their system.
}
\details{
  The primary purpose of this package is to build models using locating arrays. This can be done by using the `buildModels()` function. Please see the function documentation for examples and usage.

}
\author{
R Code Author: Leah Darwin
C++ Code Author: Stephen Seidel
Maintainer: Leah Darwin <ljdarwin@asu.edu>
}
\references{
  Original C++ code github: https://github.com/sseidel16/v4-la-tools
}
\keyword{ package }
}
